pyproject.toml
[tool.poetry]
name = "simswap"
version = "0.1.0"
description = ""
authors = ["Matheus Oliveira de Souza <msouza.os@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"


[tool.poetry.group.lib.dependencies]
opencv-python = "^4.9.0.80"
pillow = "^10.3.0"
numpy = "^1.26.4"
imageio = "^2.34.0"
moviepy = "^1.0.3"
insightface = "^0.7.3"
timm = "0.5.4"
torch = {version = "^2.2.2+cpu", source = "pytorch-cpu"}
torchvision = {version = "^0.17.2+cpu", source = "pytorch-cpu"}


[tool.poetry.group.api.dependencies]
fastapi = "^0.110.1"
uvicorn = "^0.29.0"
python-multipart = "^0.0.9"


[tool.poetry.group.demo.dependencies]
streamlit = "^1.33.0"
pillow = "^10.3.0"
httpx = "^0.27.0"
streamlit-webrtc = "^0.47.6"
mediapipe = "^0.10.11"



[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"
matplotlib = "^3.8.4"

[[tool.poetry.source]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
